# internal/httpapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loomwork HTTP API",
    "description": "JSON API the loomwork browser UI runs on, served by `loomwork serve` over loopback only. It is a handler layer over internal/store: every response document is a stored domain document, so this description doubles as the contract for scripts and agents that read a workspace without shelling out to the CLI. There is no authentication because the server is single-user, local-first, and never reachable from another host.",
    "version": "0.1.0",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    }
  },
  "security": [],
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Default `loomwork serve` address; --addr changes the port."
    }
  ],
...
  "paths": {
...
    "/api/models": {
      "get": {
        "tags": ["desktop"],
        "operationId": "listModels",
        "summary": "List configured text providers and their available models.",
        "description": "A provider discovery failure stays on that provider row and does not hide other configured providers. Configured default and preset-backed models remain selectable when live discovery is unavailable.",
        "responses": {
          "200": {
            "description": "Configured provider/model choices.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/ProviderModels"}
                }
              }
            }
          },
          "503": {"$ref": "#/components/responses/ServiceUnavailable"}
        }
      }
    },
    "/api/projects": {
      "get": {
        "tags": ["projects"],
        "operationId": "listProjects",
        "summary": "Summarize every project.",
        "description": "Each row is built from the project document alone, including the counts the store caches, so listing many projects never scans their subfolders.",
        "responses": {
          "200": {
            "description": "Project summaries.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/ProjectSummary"}
                }
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"}
        }
      },
      "post": {
        "tags": ["projects"],
        "operationId": "createProject",
        "summary": "Create a project.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CreateProjectRequest"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "The stored project document.",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Project"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"}
        }
      }
    },
    "/api/projects/{projectRef}": {
      "parameters": [{"$ref": "#/components/parameters/ProjectRef"}],
      "get": {
        "tags": ["projects"],
        "operationId": "getProject",
        "summary": "Fetch one project document.",
        "description": "Returns metadata, document sources, cached counts, and artifacts.",
        "responses": {
          "200": {
            "description": "The stored project document.",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Project"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/projects/{projectRef}/sources": {
      "parameters": [{"$ref": "#/components/parameters/ProjectRef"}],
      "get": {
        "tags": ["projects"],
        "operationId": "listSources",
        "summary": "List a project's document source links.",
        "responses": {
          "200": {"$ref": "#/components/responses/SourceList"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "post": {
        "tags": ["projects"],
        "operationId": "addSource",
        "summary": "Attach or update a document source link.",
        "description": "Posting a source whose name already exists updates that link instead of duplicating it, matching the CLI.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/DocumentSource"}
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/SourceList"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/projects/{projectRef}/requirements": {
      "parameters": [{"$ref": "#/components/parameters/ProjectRef"}],
      "get": {
        "tags": ["requirements"],
        "operationId": "listRequirements",
        "summary": "List the current version of every requirement.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Return only requirements whose current version has this status.",
            "schema": {"$ref": "#/components/schemas/RequirementStatus"}
          }
        ],
        "responses": {
          "200": {"$ref": "#/components/responses/RequirementList"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "post": {
        "tags": ["requirements"],
        "operationId": "createRequirement",
        "summary": "Create a requirement at version 1.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/RequirementWrite"}
            }
          }
        },
        "responses": {
          "201": {"$ref": "#/components/responses/RequirementVersion"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/projects/{projectRef}/requirements/{requirementId}": {
      "parameters": [
        {"$ref": "#/components/parameters/ProjectRef"},
        {"$ref": "#/components/parameters/RequirementID"}
      ],
      "get": {
        "tags": ["requirements"],
        "operationId": "getRequirement",
        "summary": "Fetch one requirement version.",
        "parameters": [{"$ref": "#/components/parameters/Version"}],
        "responses": {
          "200": {"$ref": "#/components/responses/RequirementVersion"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "put": {
        "tags": ["requirements"],
        "operationId": "amendRequirement",
        "summary": "Amend the current requirement version in place.",
        "description": "Rewrites the current version without changing its version, status, or creation timestamp. Fields left empty inherit from the current version. `status` is rejected — use the status endpoint to change a status.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/RequirementUpdate"}
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/RequirementVersion"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "patch": {
        "tags": ["requirements"],
        "operationId": "updateRequirement",
        "summary": "Write the next requirement version.",
        "description": "The new version is active and the previous one becomes superseded; earlier versions stay readable. `status` is rejected here — use the status endpoint to change a status.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/RequirementUpdate"}
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/RequirementVersion"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
...
    "/api/projects/{projectRef}/requirements/{requirementId}/status": {
      "parameters": [
        {"$ref": "#/components/parameters/ProjectRef"},
        {"$ref": "#/components/parameters/RequirementID"}
      ],
      "post": {
        "tags": ["requirements"],
        "operationId": "setRequirementStatus",
        "summary": "Set one version's status.",
        "description": "Changes a status in place rather than writing a new version, so marking a requirement obsolete does not invent history.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/StatusRequest"}
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/RequirementVersion"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/projects/{projectRef}/items": {
      "parameters": [{"$ref": "#/components/parameters/ProjectRef"}],
      "get": {
        "tags": ["desktop"],
        "operationId": "listProjectItems",
        "summary": "Build the Agent Desktop entity tree.",
        "description": "Returns requirements, agent definitions, override rules, test suites and nested cases, report files, and latest free-form artifact revisions grouped by their project subfolder.",
        "responses": {
          "200": {
            "description": "Grouped project tree.",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/TreeResponse"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/projects/{projectRef}/items/{family}/{itemRef}": {
      "parameters": [
        {"$ref": "#/components/parameters/ProjectRef"},
        {"$ref": "#/components/parameters/ItemFamily"},
        {"$ref": "#/components/parameters/ItemRef"}
      ],
      "get": {
        "tags": ["desktop"],
        "operationId": "getProjectItem",
        "summary": "Read one typed viewer document.",
        "description": "The artifactType and mediaType select the frontend viewer registry. Unknown values remain valid and use the raw fallback.",
        "parameters": [{"$ref": "#/components/parameters/Version"}],
        "responses": {
          "200": {
            "description": "Viewer registry input.",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ViewerDocument"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
...
    "/api/projects/{projectRef}/testability": {
      "parameters": [{"$ref": "#/components/parameters/ProjectRef"}],
      "get": {
        "tags": ["desktop"],
        "operationId": "getTestability",
        "summary": "Derive the project's testability rollup.",
        "description": "Computed on read from the current test suite versions and the reports folder: which active requirements have a linked test case, which have none, and the outcome of the most recent report.",
        "responses": {
          "200": {
            "description": "Coverage and last-run summary.",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/TestabilityReport"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/projects/{projectRef}/artifacts": {
      "parameters": [{"$ref": "#/components/parameters/ProjectRef"}],
      "get": {
        "tags": ["artifacts"],
        "operationId": "listArtifacts",
        "summary": "List the latest revision of every artifact name.",
        "responses": {
          "200": {
            "description": "Latest artifact revisions, ordered by name.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/Artifact"}
                }
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "post": {
        "tags": ["artifacts"],
        "operationId": "addArtifact",
        "summary": "Add an inline-content artifact revision.",
        "description": "Posting an existing name writes the next revision with the previous one as its parent, matching `artifact add`. Only inline content is accepted: a browser cannot reference a path on the server's disk.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/ArtifactWrite"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "The stored artifact revision.",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Artifact"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/projects/{projectRef}/test-suites": {
      "parameters": [{"$ref": "#/components/parameters/ProjectRef"}],
      "get": {
        "tags": ["test-suites"],
        "operationId": "listTestSuites",
        "summary": "List the current version of every test suite, manifest only.",
        "responses": {
          "200": {
            "description": "Suite manifests without their cases; case_ids names them.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {"$ref": "#/components/schemas/TestSuite"}
                }
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "post": {
        "tags": ["test-suites"],
        "operationId": "importTestSuite",
        "summary": "Import a suite document as the next version of its suite id.",
        "description": "The body is the suite document itself (docs/schemas/test-case.schema.json), so a file that works with `test-suite import --file` posts unchanged. The suite is audited against the active override rules and requirements: unlinked cases or violated rules make it incomplete, and an incomplete suite is stored and flagged rather than rejected.",
        "parameters": [
          {
            "name": "suiteId",
            "in": "query",
            "required": false,
            "description": "Override the suite_id in the document.",
            "schema": {"type": "string", "pattern": "^[a-z0-9][a-z0-9-]*$"}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/TestSuiteImport"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "The stored suite version and its audit.",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/ImportResult"}
              }
            }
          },
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/projects/{projectRef}/agent-definitions": {
      "parameters": [{"$ref": "#/components/parameters/ProjectRef"}],
      "get": {
        "tags": ["agents"],
        "operationId": "listAgentDefinitions",
        "summary": "List the current version of every agent definition.",
        "responses": {
          "200": {"$ref": "#/components/responses/AgentDefinitionList"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "post": {
        "tags": ["agents"],
        "operationId": "createAgentDefinition",
        "summary": "Create an agent definition at version 1.",
        "description": "target_provider defaults to claude-agent-sdk. tools_allowed is a strict allowlist: a definition that names no tools grants none.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/AgentDefinitionWrite"}
            }
          }
        },
        "responses": {
          "201": {"$ref": "#/components/responses/AgentDefinitionVersion"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/projects/{projectRef}/agent-definitions/{agentName}": {
      "parameters": [
        {"$ref": "#/components/parameters/ProjectRef"},
        {"$ref": "#/components/parameters/AgentName"}
      ],
      "patch": {
        "tags": ["agents"],
        "operationId": "updateAgentDefinition",
        "summary": "Write the next agent definition version.",
        "description": "Fields left empty inherit from the current version, so a body-only edit keeps the target and tool allowlist. Earlier version files are never modified.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/AgentDefinitionWrite"}
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/AgentDefinitionVersion"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/projects/{projectRef}/override-rules": {
      "parameters": [{"$ref": "#/components/parameters/ProjectRef"}],
      "get": {
        "tags": ["agents"],
        "operationId": "listOverrideRules",
        "summary": "List the current version of every override rule.",
        "responses": {
          "200": {"$ref": "#/components/responses/OverrideRuleList"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      },
      "post": {
        "tags": ["agents"],
        "operationId": "createOverrideRule",
        "summary": "Create an override rule at version 1.",
        "description": "A rule is hybrid: a structured condition/action pair the workbench audits deterministically plus the free-text rationale the agent reasons over. The rationale is required.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/OverrideRuleWrite"}
            }
          }
        },
        "responses": {
          "201": {"$ref": "#/components/responses/OverrideRuleVersion"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/projects/{projectRef}/override-rules/{ruleId}": {
      "parameters": [
        {"$ref": "#/components/parameters/ProjectRef"},
        {"$ref": "#/components/parameters/RuleID"}
      ],
      "patch": {
        "tags": ["agents"],
        "operationId": "updateOverrideRule",
        "summary": "Write the next override rule version.",
        "description": "The new version is active and the previous one becomes superseded. Fields left empty inherit from the current version.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/OverrideRuleWrite"}
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/OverrideRuleVersion"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
    "/api/projects/{projectRef}/override-rules/{ruleId}/status": {
      "parameters": [
        {"$ref": "#/components/parameters/ProjectRef"},
        {"$ref": "#/components/parameters/RuleID"}
      ],
      "post": {
        "tags": ["agents"],
        "operationId": "setOverrideRuleStatus",
        "summary": "Set one rule version's status to active or obsolete.",
        "description": "superseded is set only by writing a newer version and is rejected here.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/OverrideRuleStatusRequest"}
            }
          }
        },
        "responses": {
          "200": {"$ref": "#/components/responses/OverrideRuleVersion"},
          "400": {"$ref": "#/components/responses/BadRequest"},
          "404": {"$ref": "#/components/responses/NotFound"}
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
      "BadRequest": {
        "description": "The request was rejected: an unparseable body, an unknown field, or a value the domain refused.",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"}
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "The desktop provider service was not configured.",
        "content": {
          "application/json": {
            "schema": {"$ref": "#/components/schemas/Error"}
          }
        }
      },
...
    },
    "schemas": {
...
      "ProviderModels": {
        "type": "object",
        "required": ["provider", "kind", "status", "models"],
        "properties": {
          "provider": {"type": "string"},
          "kind": {"type": "string", "enum": ["ollama", "lmstudio", "azure", "bedrock"]},
          "status": {"type": "string"},
          "models": {
            "type": "array",
            "items": {"$ref": "#/components/schemas/ModelChoice"}
          }
        }
      },
...
    }
  }
}